chore: clean up search api route
