+++ lib/spotify.js
@@ -1,26 +1,12 @@
 /**
- * Gets query params
- */
-function _getQueryParams() {
-  return {
-    q: "track:come and play in the milky night",
-    type: "track",
-    include_external: "audio&limit=1",
-  };
-}
-
-/**
- * Gets request client
+ * Gets request init
  *
  * @param {string} accessToken
  */
-function _getRequestClient(accessToken) {
+function _getRequestInit(accessToken) {
   return {
-    method: "GET",
     headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-      "Content-Type": "application/json",
+      authorization: accessToken,
     },
   };
 }
@@ -28,16 +14,12 @@ function _getRequestClient(accessToken) {
 /**
  * Search for tracks
  *
+ * @param {string} api
  * @param {string} accessToken
- * @returns tracks
+ * @returns
  */
 export async function search(accessToken) {
-  const baseUrl = "https://api.spotify.com/";
-  const endpoint = "v1/search";
-  const params = _getQueryParams();
-  const query = new URLSearchParams(params).toString();
-  const api = baseUrl + endpoint + "?" + query;
-  const response = await fetch(api, _getRequestClient(accessToken));
-  const data = await response.json();
-  return data;
+  const api = "/api/v1/search";
+  const init = _getRequestInit(accessToken);
+  return await fetch(api, init);
 }
