+++ pages/client.js
@@ -1,5 +1,6 @@
 import Layout from "../components/layout";
 import { useRef } from "react";
+import { search } from "../lib/spotify";
 
 // Temp page to test Spotify API functionality
 // Make sure to access this from the front page
@@ -19,11 +20,7 @@ export default function GetData() {
       <button
         onClick={async () => {
           await setTokensRef();
-          const response = await fetch("/api/v1/search", {
-            headers: {
-              authorization: tokensRef.current.access_token,
-            },
-          });
+          const response = await search(tokensRef.current.access_token);
           const json = await response.json();
           console.log(json);
         }}
